04_영화 정보 및 평점 등록 웹 서비스/backend/requirements.txt
# Core
fastapi[all]==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.3.4

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3

# Security & Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# HTTP & External API
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
